Decompose docker switches when building spec
